build: create workflow to enable true `fast-forward` merge
to overcome the shortcomings of GitHubs merge strategies this
workflow seeks to enable true `fast-forward` merge of a PR through
the use of a special `fast-forward` label to a pull request.

also includes documentation for the new workflow and updates to
the `CONTRIBUTING.md` to explain how to merge pull requests.